[refactor] 패키징 & 스크린생성
