+++ app/src/main/java/com/ggaebiz/ggaebiz/di/AppModule.kt
@@ -5,7 +5,7 @@ import com.ggaebiz.ggaebiz.data.datastore.DataStoreObject.audioDataStore
 import com.ggaebiz.ggaebiz.data.repository.AudioRepositoryImpl
 import com.ggaebiz.ggaebiz.domain.repository.AudioRepository
 import com.ggaebiz.ggaebiz.domain.usecase.GetAudioResIdUseCase
-import com.ggaebiz.ggaebiz.presentation.ui.timer.TimerViewModel
+import com.ggaebiz.ggaebiz.presentation.ui.alarm.AlarmViewModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
@@ -15,5 +15,5 @@ val appModule = module {
     single { AudioDataStore(get()) }
     single<AudioRepository> { AudioRepositoryImpl(get()) }
     factory { GetAudioResIdUseCase(get()) }
-    viewModel { TimerViewModel(get()) }
+    viewModel { AlarmViewModel(get()) }
 }
