+++ app/src/main/java/com/ggaebiz/ggaebiz/presentation/common/extension/extension.kt
@@ -1,4 +1,4 @@
-package com.ggaebiz.ggaebiz.presentation.ui.timer
+package com.ggaebiz.ggaebiz.presentation.common.extension
 
 import android.content.Context
