+++ app/src/main/java/com/ggaebiz/ggaebiz/presentation/designsystem/component/header/LogoAppBar.kt
@@ -1,4 +1,4 @@
-package com.ggaebiz.ggaebiz.presentation.designsystem.component
+package com.ggaebiz.ggaebiz.presentation.designsystem.component.header
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
