+++ app/src/main/java/com/ggaebiz/ggaebiz/presentation/ui/alarm/AlarmViewModel.kt
@@ -1,4 +1,4 @@
-package com.ggaebiz.ggaebiz.presentation.ui.timer
+package com.ggaebiz.ggaebiz.presentation.ui.alarm
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -13,23 +13,23 @@ import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-data class TimerState(
+data class AlarmState(
     val text: String = "타이머를 울려주세요",
 )
 
-sealed interface TimerSideEffect {
-    data class PlayAudio(val resId: String) : TimerSideEffect
+sealed interface AlarmSideEffect {
+    data class PlayAudio(val resId: String) : AlarmSideEffect
 }
 
-class TimerViewModel(
+class AlarmViewModel(
     private val getAudioResIdUseCase: GetAudioResIdUseCase,
 ) : ViewModel() {
 
-    private val _uiState = MutableStateFlow(TimerState())
-    val uiState: StateFlow<TimerState> = _uiState.asStateFlow()
+    private val _uiState = MutableStateFlow(AlarmState())
+    val uiState: StateFlow<AlarmState> = _uiState.asStateFlow()
 
-    private val _sideEffects = Channel<TimerSideEffect>()
-    val sideEffects: Flow<TimerSideEffect> = _sideEffects.receiveAsFlow()
+    private val _sideEffects = Channel<AlarmSideEffect>()
+    val sideEffects: Flow<AlarmSideEffect> = _sideEffects.receiveAsFlow()
 
 
     fun startTimer(
@@ -38,6 +38,6 @@ class TimerViewModel(
     ) = viewModelScope.launch {
         val resIds = getAudioResIdUseCase(character, level, 0)
         _uiState.update { it.copy(text = "타이머 울리는 중") }
-        _sideEffects.send(TimerSideEffect.PlayAudio(resIds))
+        _sideEffects.send(AlarmSideEffect.PlayAudio(resIds))
     }
 }
