+++ app/src/main/java/com/ggaebiz/ggaebiz/presentation/ui/home/HomeScreen.kt
@@ -1,0 +1,9 @@
+package com.ggaebiz.ggaebiz.presentation.ui.home
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HomeScreen() {
+    Text("Home")
+}
