+++ app/src/main/java/com/ggaebiz/ggaebiz/presentation/ui/setting/SettingScreen.kt
@@ -1,0 +1,7 @@
+package com.ggaebiz.ggaebiz.presentation.ui.setting
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SettingScreen() {
+}
